backend/services/blob_service/requirements.txt
# =============================================================================
# Blob Storage Service Requirements
# =============================================================================

fastapi>=0.115.5
uvicorn[standard]>=0.32.0

# --- MinIO SDK ---------------------------------------------------------------
minio>=7.2.12                    # Official MinIO Python SDK (S3-compatible)

# --- File handling -----------------------------------------------------------
python-multipart>=0.0.20         # FastAPI multipart file upload support

# --- Database ----------------------------------------------------------------
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary

# --- Validation & Config -----------------------------------------------------
pydantic>=2.11.0
python-dotenv>=1.0.1

# --- Shared package (installed separately) -----------------------------------
# pip install -e ../../../shared
